need more config

+++ srcs/wordpress/tools/script.sh
@@ -1,6 +1,10 @@
 #! /bin/bash
 mkdir -p /run/php/
 touch /run/php/php7.3-fpm.pid
-tar -xf /root/latest.tar.gz -C /var/www/html/ --strip-components=1
-rm -rf /root/latest.tar.gz
+if test -f "/var/www/html/index.php"; then
+    echo "wp exist"
+else
+    tar -xf /root/latest.tar.gz -C /var/www/html/ --strip-components=1
+    rm -rf /root/latest.tar.gz
+fi
 $(ls /usr/sbin/ | grep php-fpm) -F
